fix: allow specifying some flags as config
API server URL, insecure mode, and CA certs can now be specified via a
config file, as well as command line arguments. This is a fix owing to
an oversight when support for config files was originally added.

As part of this, --api-server flag is no longer marked as mandatory (as
the corresponding config could be used instead).

Signed-off-by: Sergei Trofimov <sergei.trofimov@arm.com>